requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.2.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-model-utils==4.1.1
django-sortedm2m==3.0.2
django-versatileimagefield==2.0
djangorestframework==3.12.4
drf-writable-nested==0.6.3
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.9.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
python-magic==0.4.22
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rest-social-auth==6.0.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
